+++ oml/admin.py
@@ -1,0 +1,3 @@
+from .models import ModelAdminOml, StatusListFilter
+
+__all__ = ['ModelAdminOml', 'StatusListFilter']