+++ oml/test_urls.py
@@ -1,0 +1,5 @@
+from django.urls import include, path
+
+urlpatterns = [
+    path('oml/', include('oml.urls')),
+]